Fix suggestions
